refactor(test): Split tests (#331)
