+++ src/test/java/com/algolia/search/saas/AlgoliaTest.java
@@ -2,6 +2,7 @@ package com.algolia.search.saas;
 
 import org.json.JSONArray;
 import org.json.JSONException;
+import org.json.JSONObject;
 import org.junit.AfterClass;
 import org.junit.Assume;
 import org.junit.Before;
@@ -55,6 +56,80 @@ public class AlgoliaTest {
       index.clearIndex();
     } catch (AlgoliaException e) {
       //Normal
+    }
+  }
+
+  protected void waitKey(APIClient client, String key, String acls) {
+    for (int i = 0; i < 60; ++i) {
+      try {
+        JSONObject obj = client.getApiKey(key);
+        if (acls != null) {
+          if (!obj.getJSONArray("acl").toString().equals(acls)) {
+            throw new AlgoliaException("not good ACL");
+          }
+        }
+        return;
+      } catch (Exception e) {
+        // not found
+        try {
+          Thread.sleep(1000);
+        } catch (InterruptedException e1) {
+          // not fatal
+        }
+      }
+    }
+  }
+
+  protected void waitKey(Index index, String key, String acls) {
+    for (int i = 0; i < 60; ++i) {
+      try {
+        JSONObject obj = index.getApiKey(key);
+        if (acls != null) {
+          if (!obj.getJSONArray("acl").toString().equals(acls)) {
+            throw new AlgoliaException("not good ACL");
+          }
+        }
+        return;
+      } catch (Exception e) {
+        // not found
+        try {
+          Thread.sleep(1000);
+        } catch (InterruptedException e1) {
+          // not fatal
+        }
+      }
+    }
+  }
+
+  protected void waitKeyDeleted(APIClient client, String key) {
+    for (int i = 0; i < 60; ++i) {
+      try {
+        client.getApiKey(key);
+        // found
+        try {
+          Thread.sleep(1000);
+        } catch (InterruptedException e1) {
+          // not fatal
+        }
+      } catch (Exception e) {
+        return;
+      }
+    }
+  }
+
+  protected void waitKeyDeleted(Index index, String key) {
+    for (int i = 0; i < 60; ++i) {
+      try {
+        index.getApiKey(key);
+        // found
+        try {
+          Thread.sleep(1000);
+        } catch (InterruptedException e1) {
+          // not fatal
+        }
+      } catch (Exception e) {
+        return;
+      }
     }
   }
 }
